Split out CSV reading code from asset.rs
